requirements.txt
pyautogen==0.1.13
openai==0.28.1
docker==6.1.3